User updating service naive implementation & Some bug fixes.

+++ UbiBikeApp/UbiBike/app/src/main/java/pt/ulisboa/tecnico/cmu/ubibike/remote/rest/UserServiceREST.java
@@ -14,9 +14,6 @@ import retrofit2.http.POST;
 import retrofit2.http.PUT;
 import retrofit2.http.Path;
 
-/**
- * Created by André on 21-03-2016.
- */
 public interface UserServiceREST {
 
     @FormUrlEncoded
@@ -28,7 +25,7 @@ public interface UserServiceREST {
     Call<User> createUsername(@Path("username")String username,@Field("password")String password);
 
     @PUT("user/{username}/points/{points}")
-    Call<User> updateUserPoints(@Path("username")String username,@Path("points") int points);
+    Call<User> updateUserPoints(@Path("username")String username,@Path("points") long points);
 
     @GET("user/{username}")
     Call<User> getUser(@Path("username")String username);
